Modifications

+++ files/.bashrc
@@ -85,10 +85,14 @@ fi
 #if [ -f /etc/bash_completion ] && ! shopt -oq posix; then
 #    . /etc/bash_completion
 #fi
-source ~/.bin/tmuxinator.zsh
-if [ ! -f ~/timer ]; then
-    echo 'Timer!' >> ~/timer
-    bash -c "sleep 5; rm -rf ~/timer" &
-    tmux kill-session
-    tmuxinator start 
+tmux_sessions=`tmux list-sessions 2>&1`
+if `echo $tmux_sessions | grep -v "no server"`; then
+    if [ ! -f ~/timer ]; then
+        echo 'Timer!' >> ~/timer
+        bash -c "sleep 5; rm -rf ~/timer" &
+        tmux kill-session
+        tmuxinator start 
+    fi
+else
+    tmuxinator start
 fi
